# Use conditional statement instead of switch in System.Linq.Enumerable