Small improvement by using a conditional instead of switch+goto.

Baseline is c4d5f6dff367ba1197

```
PS C:\Users\acovingt\source\repos\performance> py .\scripts\benchmarks_ci.py -c Release -f net6.0 --filter LinqBenchmark* --corerun C:\Users\acovingt\source\repos\runtime-master\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts D:\dotnet-micro-results\linq-use-conditional\base --bdn-arguments "--launchCount 5"
PS C:\Users\acovingt\source\repos\performance> py .\scripts\benchmarks_ci.py -c Release -f net6.0 --filter LinqBenchmark* --corerun C:\Users\acovingt\source\repos\runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts D:\dotnet-micro-results\linq-use-conditional\diff --bdn-arguments "--launchCount 5"
PS C:\Users\acovingt\source\repos\performance\src\tools\ResultsComparer> dotnet run -- --base D:\dotnet-micro-results\linq-use-conditional\base\results\ --diff D:\dotnet-micro-results\linq-use-conditional\diff\results\ --threshold 3% --noise 5ns
summary:
better: 4, geomean: 1.057
total diff: 4

No Slower results for the provided threshold = 3% and noise filter = 5ns.

| Faster                            | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| LinqBenchmarks.Where00LinqQueryX  |      1.07 |     562294600.00 |     523471200.00 |         |
| LinqBenchmarks.Where00LinqMethodX |      1.07 |     561668600.00 |     523850750.00 |         |
| LinqBenchmarks.Where01LinqMethodX |      1.04 |     239930400.00 |     230370400.00 |         |
| LinqBenchmarks.Where01LinqQueryX  |      1.04 |     239949600.00 |     230396200.00 |         |
```