@stephentoub 2021-11-04
Thanks.

Do you have an explanation for the improvements you're seeing?  Your results cite a 7% improvement in:
https://github.com/dotnet/performance/blob/d7dac8a7ca12a28d099192f8a901cf8e30361384/src/benchmarks/micro/runtime/Linq/Linq.cs#L125-L143
which is using Where on a `List<T>` and as such shouldn't be using this code path at all.  The change you've made is to WhereEnumerableIterator but a query on a `List<T>` should be using WhereListIterator.

FWIW, I also personally find the switch format easier to read.