@stephentoub 2021-11-04
> The change you've made is to WhereEnumerableIterator but a query on a List<T> should be using WhereListIterator.

Nevermind, somehow I missed in the diff you were modifying both.