@stephentoub 2021-11-04
Even with that, I still find the switch easier to read and understand.  If we're really trying to eek out a few percentage points improvement here and willing to sacrifice understandability to do so, I'd expect it'd be even better to restructure the code so that on the hot path of _state == 2 we're not having to check _state == 1 first (_state == 1 should happen at most once for the whole enumeration).