@alexcovington 2021-11-04
@stephentoub The `goto` statement within the switch is particularly expensive in this case according to PerfView:

![linq-goto-samples](https://user-images.githubusercontent.com/68252706/140375884-e16cd816-452a-4a73-b423-a3c98714ea11.png)

> I'd expect it'd be even better to restructure the code so that on the hot path of _state == 2 we're not having to check _state == 1 first (_state == 1 should happen at most once for the whole enumeration).

I see, I did not catch that `_state == 1` only happens once. I can try to move that check out of the hot path to reduce the number of checks.