@danmoseley 2021-11-04
PerfView can give you instruction level cost? I thought it was only method level. TIL.