@alexcovington 2021-11-04
> PerfView can give you instruction level cost? I thought it was only method level. TIL.

Well, I'm not sure exactly how accurate PerfView's cost reporting here (it's just stack sampling, right?), but the benchmark is clearly spending a lot of time just doing the `goto`.