@stephentoub 2021-11-04
> but the benchmark is clearly spending a lot of time just doing the goto

I don't believe that's being correctly attributed.