@EgorBo 2021-11-04
btw, related:


```csharp
static void Test(int s)
{
    if (s == 1)
    {
        Console.WriteLine("1");
        s = 2;
    }

    if (s == 2)
    {
        Console.WriteLine("2");
    }
}
```
is optimized by jit into:
```
static void Test(int s)
{
    if (s == 1)
    {
        Console.WriteLine("1");
        s = 2;
        goto jt;
    }

    if (s == 2)
    {
jt:
        Console.WriteLine("2");
    }
}
```
🙂 (aka Jump-Threading). Not sure it kicks in here as _state is a field (but maybe still kicks in becuase of struct)