@tannergooding 2021-11-04
> There's no "goto" in the generated IL

`goto case ...` in C# is generally used as `__fallthrough` is in C/C++. C# doesn't allow implicit fallthrough and `goto case ...`/`goto default` are the alternative.
* You can use `goto case ...` to jump to a non-sequential IL statement, but that's not "super" common