@alexcovington 2021-11-04
> Not sure it kicks in here as _state is a field (but maybe still kicks in becuase of struct)

Doesn't look like it kicks in for this case:

```assembly
G_M35240_IG04:
       nop      
       mov      rcx, gword ptr [rbp+10H]
       cmp      dword ptr [rcx+20], 1         ;; Check if _state == 1
       sete     cl
       movzx    rcx, cl
       mov      dword ptr [rbp-1CH], ecx
       cmp      dword ptr [rbp-1CH], 0
       je       SHORT G_M35240_IG05						   
       nop      			  
       mov      rcx, gword ptr [rbp+10H]        ;; _state == 1
       mov      gword ptr [rbp-88H], rcx
       mov      rcx, gword ptr [rbp+10H]
       mov      rcx, gword ptr [rcx+24]
       lea      rdx, [rbp-A0H]
       cmp      dword ptr [rcx], ecx
       call     System.Collections.Generic.List`1[__Canon][System.__Canon]:GetEnumerator():Enumerator[__Canon]:this
       mov      rdi, gword ptr [rbp-88H]
       add      rdi, 40
       lea      rsi, bword ptr [rbp-A0H]
       call     CORINFO_HELP_ASSIGN_BYREF
       call     CORINFO_HELP_ASSIGN_BYREF
       movsq    
       mov      rcx, gword ptr [rbp+10H]
       mov      dword ptr [rcx+20], 2         ;; Set _state = 2, fall through
       nop      							   
						;; bbWeight=1    PerfScore 26.25
G_M35240_IG05:
       mov      rcx, gword ptr [rbp+10H]
       cmp      dword ptr [rcx+20], 2         ;; Still checks if _state == 2
       sete     cl
       movzx    rcx, cl
       mov      dword ptr [rbp-20H], ecx
       cmp      dword ptr [rbp-20H], 0
       je       G_M35240_IG15
       nop      
       nop      
       jmp      G_M35240_IG11
						;; bbWeight=1    PerfScore 11.75
```