@alexcovington 2021-11-08
@stephentoub I've been testing a few different ways to move the `_state == 1` check out of the hot path, but so far nothing has been as performant as the current PR. Adding extra logic to handle the case requires either goto's or extra comparisons, which cancel out the benefits of the PR and result in on-par performance.